+++ etils/ecolab/__init__.py
@@ -18,6 +18,7 @@ from etils.ecolab.array_as_img import auto_plot_array
 from etils.ecolab.colab_utils import collapse
 from etils.ecolab.colab_utils import interruptible
 from etils.ecolab.colab_utils import json
+from etils.ecolab.highlight_util import highlight_html
 from etils.ecolab.inspects.auto_utils import auto_inspect
 from etils.ecolab.inspects.core import inspect
 from etils.ecolab.module_utils import clear_cached_modules
@@ -25,13 +26,3 @@ from etils.ecolab.patch_utils import patch_graphviz
 from etils.ecolab.patch_utils import set_verbose
 from etils.ecolab.pyjs_com import js_import as pyjs_import
 from etils.ecolab.pyjs_com import register_js_fn
-
-__all__ = [
-    'auto_plot_array',
-    'collapse',
-    'clear_cached_modules',
-    'inspect',
-    'interruptible',
-    'patch_graphviz',
-    'set_verbose',
-]