+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-05-12
+
 *   `ecolab`:
     *   Add `widget=True` argument to `ecolab.collapse()` for better
         interactivity.
@@ -255,7 +257,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/google/etils/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/google/etils/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/google/etils/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/google/etils/compare/v1.0.0...v1.1.0